Add : 보스 데이터 저장

+++ Assets/Scripts/UI/HealthStatusUI.cs
@@ -18,7 +18,7 @@ public class HealthStatusUI : MonoBehaviour
     #region HP 업데이트
     public void UpdateHPStatus()
     {
-        HPBar.value = HealthSystem.CurHP / HealthSystem.MaxHP;
+        HPBar.value = (float)HealthSystem.CurHP / HealthSystem.MaxHP;
     }
     #endregion
     public void ShowSlider()
